+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - interface for copying dashboard link
 
+### Changed
+
+- the design of dropdown that show dashboard link
+
 ## [0.7.0]
 
 ### Added